+++ lib/bmg/sql/processor/requalify.rb
@@ -6,7 +6,7 @@ module Bmg
         def initialize(builder)
           super
           @requalify = Hash.new{|h,k|
-            h[k] = Grammar.sexpr [:range_var_name, builder.next_qualifier!]
+            h[k.to_s] = builder.next_qualifier!
           }
         end
         attr_reader :requalify 
@@ -15,7 +15,16 @@ module Bmg
         alias :on_missing :copy_and_apply
 
         def on_range_var_name(sexpr)
-          requalify[sexpr]
+          Grammar.sexpr [:range_var_name, requalify[sexpr.qualifier.to_s] ]
+        end
+
+        def on_where_clause(sexpr)
+          pred = Predicate::Grammar.sexpr(apply(sexpr.predicate))
+          sexpr([:where_clause, pred])
+        end
+
+        def on_qualified_identifier(sexpr)
+          Predicate::Factory.qualified_identifier(requalify[sexpr.qualifier.to_s].to_sym, sexpr.name.to_sym)
         end
 
       end # class Requalify